DP13: move constants and validators to service folder, change constant name
