+++ backend/services/constants.py
@@ -1,2 +1,2 @@
-MAX_NAME_LENGTH: int = 16
+MAX_USER_NAME_LENGTH: int = 16
 MAX_DEVICE_ID_LENGTH: int = 24
