+++ backend/users/models.py
@@ -1,7 +1,7 @@
-from api.constants import MAX_DEVICE_ID_LENGTH, MAX_NAME_LENGTH
-from api.validators import validate_name
 from django.contrib.auth.models import AbstractUser
 from django.db import models
+from services.constants import MAX_DEVICE_ID_LENGTH, MAX_USER_NAME_LENGTH
+from services.validators import validate_name
 
 
 class User(AbstractUser):
@@ -9,7 +9,7 @@ class User(AbstractUser):
     USERNAME_FIELD = "device_id"
     name = models.CharField(
         verbose_name="Имя",
-        max_length=MAX_NAME_LENGTH,
+        max_length=MAX_USER_NAME_LENGTH,
         validators=[validate_name],
         blank=False,
     )
